Also extract out verify-email for the purpose of stopping test bundling
